rename toolbox cmd
